Use term "LXC" in output when LXC is checked (#21)
* Use term "LXC" in output when LXC is checked

You can use the VM check for LXCs, too. Previously, the output when checking an LXC was:

./check_pve.py ... -m vm --name pihole1
OK - VM 'pihole1' on node 'proxmox3' is running|cpu=0.11%;; memory=13.52%;;

Having this patch applied, the output is changed to:

OK - LXC 'pihole1' on node 'proxmox3' is running|cpu=0%;; memory=13.5%;;

Checking a VM instead of an LXC, the output remains unchanged:

./check_pve.py ... -m vm -n proxmox3 --name icinga-alt
OK - VM 'icinga-alt' on node 'proxmox3' is running|cpu=39.71%;; memory=66.87%;;

* Remove 'named' wording

As @nbuchwitz pointed out, a VM or LXC may also be specified using a number as the identifier. For this, the wording 'named' wouldn't be appropriate.

* Docu updates for help text (LXC same as VM)

For --vmid, the container is already mentioned as possible ID.
In this commit, the --name was completed.

Also, a sample command was added.