fix(subscription): Handle all possible statuses
Maybe the most important change is the addition of an "else" branch
that reports an UNKNOWN state if none of the handled statuses matches.
Hopefully this will avoid the problem that a change of Proxmox's API
goes unnoticed, because the check still returns OK (as happened in
 https://github.com/nbuchwitz/check_pve/issues/55).

Apart from that I've added the other currently defined statuses. (As
far as I can see they are defined in
https://github.com/proxmox/proxmox-rs/blob/57b0477657/proxmox-subscription/src/subscription_info.rs#L14.)
One could argue that "suspended" should result in a CRITICAL state as
well, but WARNING also felt reasonable to me.

I'm not sure if "inactive" was a valid status at some point. Looking
through some old code in https://git.proxmox.com/?p=pve-manager.git I
didn't find a reference to that status. Maybe it was "invalid" all
the time?

Please note, that I wasn't able to test this thoroughly.